# docs/phase2/29_API_OPENAPI.yaml
openapi: 3.0.3
info:
  title: FlexSpace API
  version: "1.0.0"
  description: >
    FlexSpace Flexible Workspace Marketplace — Phase 2 API contract.
    Implements the decisions in 11_API_CONTRACTS.md. REST/JSON, versioned base path,
    JWT bearer auth, consistent error envelope, idempotency keys on mutating
    financial/booking endpoints. This is a contract, not an implementation —
    no server code exists yet.
servers:
  - url: https://api.flexspace.az/api/v1
    description: Production (placeholder domain)
  - url: https://staging-api.flexspace.az/api/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /provider/rooms:
    post:
      tags: [Provider]
      summary: Create a room (subject to plan-tier limits — 25_PROVIDER_ARCHITECTURE.md §25.2)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoomInput" }
      responses:
        "201":
          description: Room created (DRAFT)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoomDetail" }
        "402":
          description: Plan tier limit reached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /provider/rooms/{roomId}:
    patch:
      tags: [Provider]
      summary: Update a room
      parameters: [{ $ref: "#/components/parameters/RoomId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoomInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoomDetail" }
  /provider/rooms/{roomId}/status:
    patch:
      tags: [Provider]
      summary: Change room status (DRAFT/ACTIVE/INACTIVE) — additive completion (P4-2) of 09_DOMAIN_MODEL.md §9.2's Room lifecycle, which had no API surface in the original contract
      parameters: [{ $ref: "#/components/parameters/RoomId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [DRAFT, ACTIVE, INACTIVE] }
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoomDetail" }
        "403":
          description: "Cannot activate: owning provider is not VERIFIED"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /provider/rooms/{roomId}/photos:
    post:
      tags: [Provider]
      summary: Upload a room photo (18_SECURITY.md §18.4 secure upload pipeline)
      parameters: [{ $ref: "#/components/parameters/RoomId" }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Photo queued for moderation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
  /provider/rooms/{roomId}/availability-rules:
    put:
      tags: [Provider]
      summary: Set recurring availability rules for a room
      parameters: [{ $ref: "#/components/parameters/RoomId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/AvailabilityRule" }
      responses: { "200": { description: Updated } }
  /provider/rooms/{roomId}/blocked-periods:
    post:
      tags: [Provider]
      summary: Manually block a time range (05_USER_FLOWS.md §5.5)
      parameters: [{ $ref: "#/components/parameters/RoomId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startAt, endAt]
              properties:
                startAt: { type: string, format: date-time }
                endAt: { type: string, format: date-time }
                reason: { type: string }
      responses:
        "201":
          description: Blocked period created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockedPeriod" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "SLOT_UNAVAILABLE" }
            message: { type: string }
            details: { type: object }
...
    Money:
      type: object
      properties:
        amount: { type: integer, description: "minor units (qəpik)" }
        currency: { type: string, example: "AZN" }
    RoomSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        roomType: { type: string }
        providerName: { type: string }
        verified: { type: boolean }
        city: { type: string }
        district: { type: string }
        lat: { type: number }
        lng: { type: number }
        distanceKm: { type: number }
        capacityMin: { type: integer }
        capacityMax: { type: integer }
        pricePerHour: { $ref: "#/components/schemas/Money" }
        averageRating: { type: number }
        reviewCount: { type: integer }
        coverPhotoUrl: { type: string, format: uri }
        available: { type: boolean }
        relevanceScore: { type: number }
...
    RoomDetail:
      allOf:
        - $ref: "#/components/schemas/RoomSummary"
        - type: object
          properties:
            description: { type: string }
            sizeSqm: { type: number }
            # `amenities` is the room's full amenity list (including the
            # EQUIPMENT-category ones per 30_SEED_DATA.sql's taxonomy) —
            # there is no separate `equipment` field or `room.rules` TEXT
            # column anywhere in 28_DATABASE_DDL.sql or the real
            # SearchService.getRoomDetail() SQL. A prior version of this
            # schema documented both; removed as doc-only drift (found
            # while building the room detail page) rather than left to
            # mislead a future client into requesting fields that will
            # never be populated.
            amenities: { type: array, items: { type: string } }
            cancellationPolicy: { type: object }
            photos: { type: array, items: { type: string, format: uri } }
            minBookingMinutes: { type: integer }
            maxBookingMinutes: { type: integer }
    RoomInput:
      type: object
      required: [locationId, roomTypeId, name, capacityMax, basePriceAmount]
      properties:
        locationId: { type: string, format: uuid }
        roomTypeId: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        capacityMin: { type: integer }
        capacityMax: { type: integer }
        sizeSqm: { type: number }
        basePriceAmount: { type: integer }
        basePriceCurrency: { type: string, default: "AZN" }
        amenityIds: { type: array, items: { type: string, format: uuid } }
        cancellationPolicy: { type: object }
    AvailabilityRule:
      type: object
      properties:
        recurrenceType: { type: string, enum: [WEEKLY, DATE_SPECIFIC] }
        dayOfWeek: { type: integer }
        specificDate: { type: string, format: date }
        startTime: { type: string }
        endTime: { type: string }
        isOpen: { type: boolean }
    BlockedPeriod:
      type: object
      properties:
        id: { type: string, format: uuid }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        reason: { type: string }
    Photo:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        moderationStatus: { type: string }
...